itemのcomponents
